Create filter pane theme
This has now been added to Power BI as per https://docs.microsoft.com/en-us/power-bi/power-bi-report-filter-preview#theming-for-filter-pane.

It needs to be added to the Global Level in the "page" section. Take note it is customisation for both the filter pane and filter cards, even though they are different items in Power BI Desktop.